+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The 'Process' extension method for IConsumer\<TMessage\> can now be used for parallel processing of messages
+
 ## [2.4.1] - 2022-09-16
 
 ### Changed